phase 1 done

+++ tethysapp/storage_capacity_service/controllers.py
@@ -1,78 +1,144 @@
-from django.contrib.auth.decorators import login_required
-from django.http import JsonResponse
+# author: Xiaohui (Sherry) Qiao, xhqiao89@gmail.com
+
 import os
 import sys
 import csv
 from datetime import datetime
 import binascii
+import subprocess
+import tempfile
+
+from django.contrib.auth.decorators import login_required
+from django.http import JsonResponse
+
+
+# Apache should have ownership and full permission over this path
+DEM_FULL_PATH = "/home/drew/Downloads/BearCk.tif"
+DEM_NAME = 'BearCk' # DEM layer name, no extension (no .tif)
+GISBASE = "/usr/lib/grass70" # the full path to GRASS installation
+GRASS7BIN = "grass70" # the command to start GRASS from shell
 
 @login_required()
 def home(request):
     """
     Controller for the app home page.
     """
-    #http://127.0.0.1:8000/apps/sc-service/?x=1696907&y=7339134.4&damh=200.0&interval=15.0
-    output_csv = "sc.csv"
+    # http://127.0.0.1:8000/apps/storage-capacity-service/?xlon=1696907&ylat=7339134.4&prj=native&damh=200.0&interval=15.0
+    # http://127.0.0.1:8000/apps/storage-capacity-service/?xlon=1696107&ylat=7339934.4&prj=native&damh=200.0&interval=5.0
 
+    # output_csv = "sc.csv"
+
+    string_length = 4
+    jobid = binascii.hexlify(os.urandom(string_length))
     time_start = datetime.now()
     status = "error"
-    message = "message"
+    message = ""
     sc = []
+    input_para = {}
 
     try:
         if request.GET:
-            x = request.GET["x"]
-            y = request.GET["y"]
-            damh = request.GET["damh"]
-            interval = request.GET["interval"]
+            xlon = request.GET.get("xlon", None)
+            ylat = request.GET.get("ylat", None)
+            prj = request.GET.get("prj", None)
+            damh = request.GET.get("damh", None)
+            interval = request.GET.get("interval", None)
+            if xlon is None or ylat is None or prj is None or damh is None or interval is None:
+                raise Exception("Valid query is like: http://.../?xlon=1696907&ylat=7339134.4&prj=native&damh=200.0&interval=15.0")
+
+            input_para = {}
+            input_para["xlon"] = xlon
+            input_para["ylat"] = ylat
+            input_para["prj"] = prj
+            input_para["damh"] = damh
+            input_para["interval"] = interval
+
             interval = abs(float(interval))
             if interval == 0:
                 interval = 10
                 message = "interval should be greater than 0! reset to {}".format(interval)
-            sc_list = SC(x, y, damh, interval)
+            sc_list, msg = SC(jobid, xlon, ylat, prj, damh, interval)
 
             if sc_list is not None:
                 status = "success"
                 sc = sc_list
-                with open(output_csv, 'wb') as fcsv:
-                    writer = csv.writer(fcsv)
-                    writer.writerows(sc_list)
+                message += msg
+                # with open(output_csv, 'wb') as fcsv:
+                #     writer = csv.writer(fcsv)
+                #     writer.writerows(sc_list)
             else:
-                status = "success"
+                status = "error"
                 sc = []
+                message += msg
+        else:
+            raise Exception("Pleas call this service in a GET request.")
     except Exception as ex:
         status = "error"
         message = ex.message
     finally:
         result_dict = {}
+        result_dict['job_id'] = jobid
         result_dict['t_start'] = time_start
         result_dict['t_end'] = datetime.now()
         result_dict["status"] = status
         result_dict["message"] = message
         result_dict['sc'] = sc
+        result_dict["input"] = input_para
+        result_dict["Note"] = "If you encountered strange errors. Please contact Admin with the JobID."
         return JsonResponse(result_dict)
 
 
-def SC(x, y, damh, interval):
-    dem = 'BearCk'
-    gisdb = "/home/drew/grass_folder"
-    location = "newLocation"
-    mapset = "drew"
+def SC(jobid, xlon, ylat, prj, damh, interval):
+
+    dem_full_path = DEM_FULL_PATH
+    dem = DEM_NAME
+    gisbase = GISBASE
+    grass7bin = GRASS7BIN
+
+    gisdb = os.path.join(tempfile.gettempdir(), 'grassdata')
+    location = "sc-location"
+    mapset = "PERMANENT"
+    keep_intermediate = False
+    msg = ""
+
+    log_name = 'log_{0}.log'.format(jobid)
+    log_path = os.path.join(gisdb, log_name)
+
+    f = open(log_path, 'w', 0)
 
     temp_files_list = []
     try:
-        string_length = 4
-        jobid = binascii.hexlify(os.urandom(string_length))
+        # override for now with TEMP dir
+        if not os.path.exists(gisdb):
+            f.write('\n---------create GISDB--------------------\n')
+            f.write('{0}\n'.format(gisdb))
+            os.mkdir(gisdb)
 
-        outlet = (float(x), float(y))
+        location_path = os.path.join(gisdb, location)
+        if not os.path.exists(location_path):
+            f.write('\n---------create Location from DEM--------------------\n')
+            f.write('{0}\n'.format(location_path))
+            startcmd = grass7bin + ' -c ' + dem_full_path + ' -e ' + location_path
+
+            print startcmd
+            p = subprocess.Popen(startcmd, shell=True,
+                                 stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            out, err = p.communicate()
+            if p.returncode != 0:
+                print >>sys.stderr, 'ERROR: %s' % err
+                print >>sys.stderr, 'ERROR: Cannot generate location (%s)' % startcmd
+                f.write('\n---------create Location failed--------------------\n')
+                sys.exit(-1)
+            else:
+                f.write('\n---------create Location done--------------------\n')
+                print 'Created location %s' % location_path
+        #---------------------
+
+        xlon = float(xlon)
+        ylat = float(ylat)
+        outlet = (xlon, ylat)
         dam_h = float(damh)
         elev_interval = float(interval)
-
-        log_name = 'log_{0}.log'.format(jobid)
-        f = open(log_name, 'w', 0)
-
-        gisbase = "/usr/lib/grass70"
-
 
         # Set GISBASE environment variable
         os.environ['GISBASE'] = gisbase
@@ -102,6 +168,11 @@ def SC(x, y, damh, interval):
         gsetup.init(gisbase, gisdb, location, mapset)
         f.write(str(gscript.gisenv()))
 
+        dem_in_mapset_path = location_path = os.path.join(gisdb, location, mapset, "cell", dem)
+        if not os.path.exists(dem_in_mapset_path):
+            f.write("\n ---------- import DEM file ------------- \n")
+            stats = gscript.read_command('r.in.gdal', input=dem_full_path, output=dem)
+
         f.write("\n ---------- raster ------------- \n")
         for rast in gscript.list_strings(type='rast'):
             f.write(str(rast))
@@ -109,14 +180,38 @@ def SC(x, y, damh, interval):
         for vect in gscript.list_strings(type='vect'):
             f.write(str(vect))
 
-        f.write("\n ---------- REAL JOB STARTS HERE ------------- \n")
-        #User input
-        #Units all in meter
+        f.write("\n ---------- !!!!!!!!!!!  REAL JOB STARTS HERE !!!!!!!!------------- \n")
 
+        # project xlon, ylat wgs84 into current
+        if prj.lower() != "native" or prj.lower() == "wgs84":
+            f.write("\n ---------- Reproject xlon and ylat into native dem projection ------------- \n")
+            stats = gscript.read_command('m.proj', coordinates=(xlon, ylat), flags='i')
+            coor_list = stats.split("|")
+            xlon = float(coor_list[0])
+            ylat = float(coor_list[1])
+            outlet = (xlon, ylat)
 
         # Define region
         f.write("\n ---------- Define region ------------- \n")
-        stats = gscript.read_command('g.region', raster=dem, flags='p')
+        stats = gscript.parse_command('g.region', raster=dem, flags='p')
+        f.write(str(stats))
+        for key in stats:
+            if "north:" in key:
+                north = float(key.split(":")[1])
+            elif "south:" in key:
+                south = float(key.split(":")[1])
+            elif "west:" in key:
+                west = float(key.split(":")[1])
+            elif "east:" in key:
+                east = float(key.split(":")[1])
+
+        # check if xlon, ylat is within the extent of dem
+        if xlon < west or xlon > east:
+            f.write("\n ERROR: xlon is out of dem region. \n")
+            raise Exception("(xlon, ylat) is out of dem region.")
+        elif ylat < south or ylat > north:
+            f.write("\n ERROR: ylat is out of dem region. \n")
+            raise Exception("(xlon, ylat) is out of dem region.")
 
         # Flow accumulation analysis
 
@@ -184,7 +279,6 @@ def SC(x, y, damh, interval):
             print("No. {0}--------------------> sc is {1} \n".format(count, str(storage)))
             storage_list.append(storage)
 
-
         for sc in storage_list:
             f.write(str(sc))
             f.write("\n")
@@ -192,16 +286,20 @@ def SC(x, y, damh, interval):
         f.write(str(datetime.now()))
         f.close()
         print ("---------------------------end--------------------------------")
-        return storage_list
+        keep_intermediate = False
+        return storage_list, msg
     except Exception as e:
+        keep_intermediate = True
         print e.message
+        msg = e.message
         if f is not None:
             f.write("\n!!!!!!  ERROR  !!!!!!\n")
             f.write(e.message)
             f.close()
-        return None
+        return None, msg
     finally:
-        for f in temp_files_list:
-            f_fullpath = "{0}/{1}/{2}/cell/{3}".format(gisdb, location, mapset, f)
-            if os.path.exists(f_fullpath):
-                os.remove(f_fullpath)
+        if not keep_intermediate:
+            for f in temp_files_list:
+                f_fullpath = "{0}/{1}/{2}/cell/{3}".format(gisdb, location, mapset, f)
+                if os.path.exists(f_fullpath):
+                    os.remove(f_fullpath)
